fs: git: move isfile/walk to fsspec wrapper
